% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mregress_crime.R
\name{mregress_crime}
\alias{mregress_crime}
\title{multiple regression}
\usage{
mregress_crime()
}
\value{
this function returns the multiple regression between all variables
}
\description{
multiple regression
}
