% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr2plot.R
\name{tr2plot}
\alias{tr2plot}
\title{tr2plot}
\usage{
tr2plot()
}
\value{
The output shows a graph built using the "ggplot2 package"
}
\description{
tr2plot
}
\author{
Elisa Boccini \email{elisa.boccini@gmail.com}
}
