% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trplot.R
\name{trplot}
\alias{trplot}
\title{trplot}
\usage{
trplot()
}
\value{
The output shows a graph built using the "ggplot2 package"
}
\description{
The "trplot" function shows how the crime rate is positively correlated to the Trump voters.
while the points are the more darker, the more we move to the right, which means the higher number of Trump voters among lower income households.
}
\author{
Elisa Boccini \email{elisa.boccini@gmail.com}
}
