% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trump}
\alias{trump}
\title{Dataset of crime in USA and its dependencies by some variables such as key socioeconomic
factors, including indicators for education (percent of adults 25 and older with at least
a high school degree, as of 2009), diversity (percent nonwhite population and percent
noncitizen population, 2015), geographic heterogeneity (percent population in metropolitan
areas, 2015), economic health (median household income, 2016 seasonally adjusted
unemployment, September 2016, percent poverty among white people. 2015, and
income inequality as measured by the Gini index, 2015), and what percent of the
population voted for Donald Trump.}
\format{An object of class \code{data.frame} with 51 rows and 12 columns.}
\source{
\url{https://github.com/fivethirtyeight/data/tree/master/hate-crimes}
}
\usage{
trump
}
\description{
#' @format A data frame with 51 rows and 12 variables:
\describe{
  \item{state}{USA States names}
  \item{median_household_income}{Median household income, 2016}
  \item{share_unemployed_seasonal}{Share of the population that is unemployed (seasonally adjusted)}
  \item{share_population_in_metro_areas}{Share of the population that lives in metropolitan areas}
  \item{share_population_with_high_school_degree}{Share of adults 25 and older with a high-school degree}
  \item{share_non_citizen}{Share of the population that are not U.S. citizens}
  \item{share_white_poverty}{Share of white residents who are living in poverty}
  \item{gini_index}{Gini Index, 2015}
  \item{share_non_white}{Share of the population that is not white}
  \item{share_voters_voted_trump}{Share of 2016 U.S. presidential voters who voted for Donald Trump}
  \item{hate_crimes_per_100k_splc}{Hate crimes per 100,000 population, Southern Poverty Law Center, Nov. 9-18, 2016}
  \item{avg_hatecrimes_per_100k_fbi}{Average annual hate crimes per 100,000 population, FBI, 2010-2015}
  ...
}
}
\keyword{datasets}
